Add embedded resources utilities for HttpGenerator.Core.Tests
This commit adds a new utility class, EmbeddedResources, in HttpGenerator.Core.Tests.Resources namespace to fetch embedded resources such as swagger petstore's json and yaml data for both V2 and V3 versions. This will be primarily used in test cases where needing to mock reading data from these resources.

A new enum class SampleOpenSpecifications is also added as part of this commit, which will be used to select the specific resource version that we want to fetch using the new utility class.