Add OpenApiDocumentFactory for file/url handling
A new class, OpenApiDocumentFactory, has been introduced that enables file and URL handling supports for OpenAPI documents. This class can asynchronously create an instance of OpenApiDocument from either a specified path or URL. It recognizes and processes both JSON and YAML formats. The path or URL can be an HTTP URL, a local YAML file, or a JSON file. Moreover, the GetHttpContent method fetches and decompresses HTTP content if necessary. The IsHttp and IsYaml methods are helper functions to determine if a given string is an HTTP URL or a YAML file respectively.