Add GeneratorSettings in HttpGenerator.Core
Introduced a new class, GeneratorSettings, in HttpGenerator.Core which provides settings for the .http file generator. This class includes a constant for default output folder and a property to fetch or set the path to the Open API. This addition will allow more configurable and versatile utilisation of the .http file generator.