Add StringCasingExtensions to handle various case conversions
A new file "StringCasingExtensions.cs" has been added to the HttpGenerator.Core. It includes methods for converting string cases such as KebabCase to PascalCase, Route to CamelCase, and spaces to PascalCase, as well as converting the first character to uppercase for the case conversions. This mitigates manual case changing and promotes code reusability.