Add option to skip validation in HttpGenerator
In the HttpGenerator, a new configuration setting `SkipValidation` has been added. Previously, the `ValidateOpenApiSpec` was always called which validated the OpenAPI specification file each time. However, with this change, users are now able to skip this validation step if necessary by setting `SkipValidation` to true. This option could potentially save time in cases where the validation is known to be unnecessary or has already been performed.