Add ContentType and BaseUrl to HttpGenerator settings
In this commit, two properties have been added to the HttpGenerator settings: ContentType and BaseUrl. ContentType serves as a default Content-Type header for all requests, while BaseUrl is a default Base URL for the same. This change provides the user with the ability to specify or override the Content-Type and Base URL for all requests generated by HttpGenerator, offering additional flexibility and control. It's especially handy if the OpenAPI specification being used doesn't specify these parameters.