Add option for generating single or multiple .http files
This commit introduces an option in the `HttpFileGenerator` for generating either a single HTTP file or multiple HTTP files based on the `OutputType` in `GeneratorSettings`. The `System.Diagnostics` import was removed as it wasn't being used. The code for generating the requests was moved to a separate `GenerateRequest` function to avoid repetition. These changes improve the code organization and offer more flexibility for the users.