Add new output type option to HttpGenerator
Added a new command line option `--output-type` to HttpGenerator in the README.md and src/HttpGenerator/README.md files. This allows the user to specify the output type for the generated HTTP request files (`OneRequestPerFile` or `OneFile`). This gives additional flexibility to the users when they want to generate HTTP requests from an OpenAPI specification.