Refactor HttpFileGenerator, remove unnecessary Json Example usage
We've refactored the HttpFileGenerator class to simplify the code and improve readability. The usage of System.Text.Json was removed as it seems unnecessary and has been replaced with an alternative method to render sample Json. Also, we created a direct call to the OperationNameGenerator in generating the operation name, this reduced complexity and overall redundancy.