Add unsafe server certificate validation bypass
The HttpClientHandlers in both the OpenApiDocumentFactory and the OpenApiValidator have been updated to skip server certificate validation. While this is generally considered a security risk, it was necessary in this context to allow for communication with certain servers that use self-signed certificates. Future iterations should aim to implement a more secure method of allowing these calls.