Use the Authority part of OpenAPI path as the Base URL if not specified
Several tests were added to validate the handling of URLs in the HttpFileGenerator. These tests check that files are generated correctly for both http and https URLs using different output types. Additionally, modified the HttpFileGenerator to correct the base URL when necessary. If the base URL is not a well-formed URL, and if the 'OpenApiPath' starts with "http", the base URL is updated to use the authority detail from the 'OpenApiPath', ensuring the correct URL is used for API calls.

Updated the SwaggerPetstore.json and SwaggerPetstore.yaml specification files with the latest version of the spec from https://petstore3.swagger.io