Remove fail-fast option from smoke tests workflow
The `fail-fast: false` option was removed from the 'file' and 'url' workflows in the smoke-tests.yml GitHub Actions file. This change was entailed due to the increased necessity for debuggable and accurate information in the event of test failures. Without the `fail-fast` option, all jobs will run to completion, making it easier to diagnose issues when there are multiple failures.