Update smoke-tests script for better output handling [skip ci]
Updated the `smoke-tests.ps1` script to include a new mandatory output parameter. The HTTP files are also removed after the generation of the new files in the `Generated` folder. This change is to handle the output files more effectively and provide more detailed command inputs for the HttpGenerator. It also allows for a cleaner workspace by removing old HTTP files.